lmn-authority-api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lmn-authority-api"
dynamic = ["version"]
description = "LMN Authority API — single source of truth for LINBO host & config data"
requires-python = ">=3.11"
license = "AGPL-3.0-or-later"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "pydantic>=2.5,<3",
    "pydantic-settings>=2.1,<3",
    "aiosqlite>=0.20,<1",
    "watchfiles>=0.21,<1",
    "pyyaml>=6.0,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "httpx>=0.27,<1",
    "pytest-cov>=5,<6",
    "ruff>=0.4,<1",
    "openapi-spec-validator>=0.7,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["lmn_authority"]

[tool.hatch.version]
path = "VERSION"
pattern = "(?P<version>.+)"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.coverage.run]
source = ["lmn_authority"]
omit = ["tests/*"]
